ETS Autopilot 0.3.2
Hey! i've posted a new version of ETS Autopilot project, here are the first ever video of working on it with debug lines.

WARNING! This version can be unstable and can make on your computer an Defender issues so we apologize to turn off Defender before starting up Autopilot executer.

We also giving to you an video to prove that it works

https://drive.google.com/file/d/1dxOBdpsnVLoseLlbH27PUjGz7Hth-GwG/view?usp=sharing